+++ src/main/java/dk/ek/persistence/CarportMapper.java
@@ -1,9 +1,8 @@
-    package dk.ek.services;
+    package dk.ek.persistence;
 
     import dk.ek.entities.Carport;
     import dk.ek.entities.Material;
     import dk.ek.entities.RoofType;
-    import dk.ek.persistence.ConnectionPool;
 
     import java.sql.Connection;
     import java.sql.PreparedStatement;
@@ -12,10 +11,10 @@
     import java.util.ArrayList;
     import java.util.List;
 
-    public class CarportService {
+    public class CarportMapper {
         private ConnectionPool connectionPool;
 
-        public CarportService(ConnectionPool connectionPool) {
+        public CarportMapper(ConnectionPool connectionPool) {
             this.connectionPool = connectionPool;
         }
